bin system #264

+++ backend/masz/Controllers/api/v1/ModCaseController.cs
@@ -90,7 +90,7 @@ namespace masz.Controllers
         }
 
         [HttpDelete("{modcaseid}")]
-        public async Task<IActionResult> DeleteSpecificItem([FromRoute] string guildid, [FromRoute] string modcaseid, [FromQuery] bool sendNotification = true, [FromQuery] bool handlePunishment = true, [FromQuery] bool announceDm = true) 
+        public async Task<IActionResult> DeleteSpecificItem([FromRoute] string guildid, [FromRoute] string modcaseid, [FromQuery] bool sendNotification = true, [FromQuery] bool handlePunishment = true, [FromQuery] bool announceDm = true, [FromQuery] bool forceDelete = false) 
         {
             logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | Incoming request.");
             Identity currentIdentity = await identityManager.GetIdentity(HttpContext);
@@ -120,14 +120,25 @@ namespace masz.Controllers
                 return NotFound();
             }
 
-            try {
-                filesHandler.DeleteDirectory(Path.Combine(config.Value.AbsolutePathToFileUpload, guildid, modcaseid));
-            } catch (Exception e) {
-                logger.LogError(e, "Failed to delete files directory for modcase.");
-            }
+            if (forceDelete && config.Value.SiteAdminDiscordUserIds.Contains(currentUser.Id))  // only siteadmin should be able to force delete.
+            {
+                try {
+                    filesHandler.DeleteDirectory(Path.Combine(config.Value.AbsolutePathToFileUpload, guildid, modcaseid));
+                } catch (Exception e) {
+                    logger.LogError(e, "Failed to delete files directory for modcase.");
+                }
 
-            database.DeleteSpecificModCase(modCase);
-            await database.SaveChangesAsync();
+                logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | Force deleting ModCase.");
+                database.DeleteSpecificModCase(modCase);
+                await database.SaveChangesAsync();
+            } else {
+                modCase.MarkedToDeleteAt = DateTime.UtcNow.AddDays(7);
+                modCase.DeletedByUserId = currentUser.Id;
+
+                logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | Marking modcase as deleted.");
+                database.UpdateModCase(modCase);
+                await database.SaveChangesAsync();
+            }
 
             if (handlePunishment)
             {
@@ -183,6 +194,12 @@ namespace masz.Controllers
                 logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | 404 ModCase not found.");
                 return NotFound();
             }
+            if (modCase.MarkedToDeleteAt != null)
+            {
+                logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | 400 Case is marked to be deleted.");
+                return BadRequest("Case is marked to be deleted.");
+            }
+
             ModCase oldModCase = (ModCase) modCase.Clone();
 
             modCase.Title = newValue.Title;
@@ -507,6 +524,12 @@ namespace masz.Controllers
                 return BadRequest("Comments are already locked.");
             }
 
+            if (modCase.MarkedToDeleteAt != null)
+            {
+                logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | 400 Case is marked to be deleted.");
+                return BadRequest("Case is marked to be deleted.");
+            }
+
             modCase.AllowComments = false;
             modCase.LockedAt = DateTime.Now;
             modCase.LockedByUserId = currentUser.Id;
@@ -559,6 +582,12 @@ namespace masz.Controllers
             if (modCase.AllowComments) {
                 logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | 400 Comments are already unlocked.");
                 return BadRequest("Comments are already unlocked.");
+            }
+
+            if (modCase.MarkedToDeleteAt != null)
+            {
+                logger.LogInformation($"{HttpContext.Request.Method} {HttpContext.Request.Path} | 400 Case is marked to be deleted.");
+                return BadRequest("Case is marked to be deleted.");
             }
 
             modCase.AllowComments = true;
